HDDS-6624. Introduce BlockLocationInfo as parent of OmKeyLocationInfo
